revert get_matrix_and_metadata to get_matrix_and_labels
get_matrix_and_labels once again returns two items, the matrix and its
labels. The AssocSpace.from_matrix factory does not accept the matrix
sum as a parameter. This wouldn't have saved much time if it worked,
because it has to sum the rows of the matrix anyway.

Also, a code review fix: slicing an AssocSpace with [:] returns a copy
of the AssocSpace.